:recycle: api: migrate session creation to new enrollment context
