+++ api/lib/application/sessions/session-controller.js
@@ -1,10 +1,10 @@
 import lodash from 'lodash';
 
+import * as certificationCandidateSerializer from '../../../src/certification/enrolment/infrastructure/serializers/certification-candidate-serializer.js';
+import * as sessionSerializer from '../../../src/certification/enrolment/infrastructure/serializers/session-serializer.js';
 import { usecases as sessionUsecases } from '../../../src/certification/session/domain/usecases/index.js';
 import * as sessionValidator from '../../../src/certification/session/domain/validators/session-validator.js';
 import * as jurySessionRepository from '../../../src/certification/session/infrastructure/repositories/jury-session-repository.js';
-import * as certificationCandidateSerializer from '../../../src/certification/session/infrastructure/serializers/jsonapi/certification-candidate-serializer.js';
-import * as sessionSerializer from '../../../src/certification/session/infrastructure/serializers/jsonapi/session-serializer.js';
 import { tokenService } from '../../../src/shared/domain/services/token-service.js';
 import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
 import * as queryParamsUtils from '../../../src/shared/infrastructure/utils/query-params-utils.js';
