+++ api/src/certification/enrolment/routes.js
@@ -1,0 +1,5 @@
+import * as session from './application/session-route.js';
+
+const certificationEnrolmentRoutes = [session];
+
+export { certificationEnrolmentRoutes };
