+++ api/src/certification/session/application/http-error-mapper-configuration.js
@@ -1,7 +1,6 @@
 import { HttpErrors } from '../../../shared/application/http-errors.js';
 import { DomainErrorMappingConfiguration } from '../../../shared/application/models/domain-error-mapping-configuration.js';
 import {
-  CertificationCandidateForbiddenDeletionError,
   SessionAlreadyFinalizedError,
   SessionAlreadyPublishedError,
   SessionWithAbortReasonOnCompletedCertificationCourseError,
@@ -29,10 +28,6 @@ const sessionDomainErrorMappingConfiguration = [
   {
     name: SessionAlreadyPublishedError.name,
     httpErrorFn: (error) => new HttpErrors.BadRequestError(error.message, error.code),
-  },
-  {
-    name: CertificationCandidateForbiddenDeletionError.name,
-    httpErrorFn: (error) => new HttpErrors.ForbiddenError(error.message, error.code),
   },
 ].map((domainErrorMappingConfiguration) => new DomainErrorMappingConfiguration(domainErrorMappingConfiguration));
