+++ api/src/certification/session/application/session-route.js
@@ -1,59 +1,11 @@
 import Joi from 'joi';
 
 import { authorization } from '../../../../lib/application/preHandlers/authorization.js';
-import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { sessionController } from './session-controller.js';
 
 const register = async function (server) {
   server.route([
-    {
-      method: 'POST',
-      path: '/api/certification-centers/{certificationCenterId}/session',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkUserIsMemberOfCertificationCenter,
-            assign: 'isMemberOfCertificationCenter',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            certificationCenterId: identifiersType.certificationCenterId,
-          }),
-        },
-        handler: sessionController.createSession,
-        tags: ['api', 'certification-center', 'sessions'],
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            '- Elle permet de créer une session de certification liée au centre de certification de l’utilisateur',
-        ],
-      },
-    },
-    {
-      method: 'PATCH',
-      path: '/api/sessions/{id}',
-      config: {
-        validate: {
-          params: Joi.object({
-            id: identifiersType.sessionId,
-          }),
-        },
-        pre: [
-          {
-            method: authorization.verifySessionAuthorization,
-            assign: 'authorizationCheck',
-          },
-        ],
-        handler: sessionController.update,
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            "- Modification d'une session de certification\n" +
-            '- L‘utilisateur doit avoir les droits d‘accès à l‘organisation liée à la session à modifier',
-        ],
-        tags: ['api', 'session'],
-      },
-    },
     {
       method: 'PUT',
       path: '/api/sessions/{id}/finalization',
@@ -87,29 +39,6 @@ const register = async function (server) {
           '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
             '- Elle permet de finaliser une session de certification afin de la signaler comme terminée',
         ],
-      },
-    },
-    {
-      method: 'DELETE',
-      path: '/api/sessions/{id}',
-      config: {
-        validate: {
-          params: Joi.object({
-            id: identifiersType.sessionId,
-          }),
-        },
-        pre: [
-          {
-            method: authorization.verifySessionAuthorization,
-            assign: 'authorizationCheck',
-          },
-        ],
-        handler: sessionController.remove,
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès au centre de certification**\n" +
-            "- Supprime la session et les candidats si la session n'a pas démarrée",
-        ],
-        tags: ['api', 'session'],
       },
     },
   ]);
