+++ api/src/certification/shared/application/http-error-mapper-configuration.js
@@ -1,6 +1,7 @@
 import { HttpErrors } from '../../../shared/application/http-errors.js';
 import { DomainErrorMappingConfiguration } from '../../../shared/application/models/domain-error-mapping-configuration.js';
 import { courseDomainErrorMappingConfiguration } from '../../course/application/http-error-mapper-configuration.js';
+import { enrolmentDomainErrorMappingConfiguration } from '../../enrolment/application/http-error-mapper-configuration.js';
 import { sessionDomainErrorMappingConfiguration } from '../../session/application/http-error-mapper-configuration.js';
 import { CertificationCourseUpdateError, InvalidCertificationReportForFinalization } from '../domain/errors.js';
 
@@ -19,6 +20,7 @@ const certificationDomainErrorMappingConfiguration = [
 
 certificationDomainErrorMappingConfiguration.push(
   ...courseDomainErrorMappingConfiguration,
+  ...enrolmentDomainErrorMappingConfiguration,
   ...sessionDomainErrorMappingConfiguration,
 );
 export { certificationDomainErrorMappingConfiguration };
