+++ api/tests/certification/enrolment/acceptance/application/session-route_test.js
@@ -1,0 +1,184 @@
+import {
+  createServer,
+  databaseBuilder,
+  expect,
+  generateValidRequestAuthorizationHeader,
+  knex,
+} from '../../../../test-helper.js';
+
+describe('Acceptance | Controller | Session | session-route', function () {
+  let server;
+
+  beforeEach(async function () {
+    server = await createServer();
+  });
+
+  describe('POST /certification-centers/{certificationCenterId}/session', function () {
+    let options;
+
+    beforeEach(function () {
+      const userId = databaseBuilder.factory.buildUser().id;
+      const certificationCenterId = databaseBuilder.factory.buildCertificationCenter({ name: 'Tour Gamma' }).id;
+      databaseBuilder.factory.buildCertificationCenterMembership({ userId, certificationCenterId });
+      options = {
+        method: 'POST',
+        url: `/api/certification-centers/${certificationCenterId}/session`,
+        payload: {
+          data: {
+            type: 'sessions',
+            attributes: {
+              'certification-center-id': certificationCenterId,
+              address: 'Nice',
+              date: '2017-12-08',
+              description: '',
+              examiner: 'Michel Essentiel',
+              room: '28D',
+              time: '14:30',
+            },
+          },
+        },
+        headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
+      };
+      return databaseBuilder.commit();
+    });
+
+    it('should return an OK status after saving in database', async function () {
+      // when
+      const response = await server.inject(options);
+
+      // then
+      const sessions = await knex('sessions').select();
+      expect(response.statusCode).to.equal(200);
+      expect(sessions).to.have.lengthOf(1);
+    });
+
+    describe('Resource access management', function () {
+      it('should respond with a 401 - unauthorized access - if user is not authenticated', async function () {
+        // given
+        options.headers.authorization = 'invalid.access.token';
+
+        // when
+        const response = await server.inject(options);
+
+        // then
+        expect(response.statusCode).to.equal(401);
+      });
+    });
+  });
+
+  describe('PATCH /api/sessions/{id}', function () {
+    let user, unauthorizedUser, certificationCenter, session, payload;
+
+    beforeEach(async function () {
+      user = databaseBuilder.factory.buildUser();
+      unauthorizedUser = databaseBuilder.factory.buildUser();
+      certificationCenter = databaseBuilder.factory.buildCertificationCenter();
+      databaseBuilder.factory.buildCertificationCenterMembership({
+        userId: user.id,
+        certificationCenterId: certificationCenter.id,
+      });
+      session = databaseBuilder.factory.buildSession({
+        certificationCenter: certificationCenter.name,
+        certificationCenterId: certificationCenter.id,
+        address: 'Nice',
+        room: '28D',
+        examiner: 'Antoine Toutvenant',
+        date: '2017-12-08',
+        time: '14:30',
+        description: 'ahah',
+        accessCode: 'ABCD12',
+      });
+      databaseBuilder.factory.buildCertificationCourse({
+        sessionId: session.id,
+      });
+      payload = {
+        data: {
+          id: session.id,
+          type: 'sessions',
+          attributes: {
+            address: 'New address',
+            room: 'New room',
+            examiner: 'Antoine Toutvenant',
+            date: '2017-08-12',
+            time: '14:30',
+            description: 'ahah',
+            accessCode: 'ABCD12',
+          },
+        },
+      };
+
+      await databaseBuilder.commit();
+    });
+
+    it('should respond with a 200 and update the session', function () {
+      const options = {
+        method: 'PATCH',
+        url: `/api/sessions/${session.id}`,
+        headers: { authorization: generateValidRequestAuthorizationHeader(user.id) },
+        payload,
+      };
+
+      // when
+      const promise = server.inject(options);
+
+      // then
+      return promise.then((response) => {
+        expect(response.statusCode).to.equal(200);
+        expect(response.result.data.type).to.equal('sessions');
+        expect(response.result.data.id).to.equal(session.id.toString());
+        expect(response.result.data.attributes.address).to.equal('New address');
+        expect(response.result.data.attributes.room).to.equal('New room');
+      });
+    });
+
+    it('should respond with a 404 when user is not authorized to update the session (to keep opacity on whether forbidden or not found)', function () {
+      const options = {
+        method: 'PATCH',
+        url: `/api/sessions/${session.id}`,
+        headers: { authorization: generateValidRequestAuthorizationHeader(unauthorizedUser.id) },
+        payload,
+      };
+
+      // when
+      const promise = server.inject(options);
+
+      // then
+      return promise.then((response) => {
+        expect(response.statusCode).to.equal(404);
+      });
+    });
+  });
+
+  describe('DELETE /sessions/{id}', function () {
+    it('should respond with 204', async function () {
+      // given
+      const server = await createServer();
+      const userId = databaseBuilder.factory.buildUser().id;
+
+      const { id: certificationCenterId, name: certificationCenter } =
+        databaseBuilder.factory.buildCertificationCenter();
+
+      const sessionId = databaseBuilder.factory.buildSession({ certificationCenterId, certificationCenter }).id;
+      databaseBuilder.factory.buildCertificationCenterMembership({ userId, certificationCenterId });
+
+      await databaseBuilder.commit();
+      const options = {
+        headers: {
+          authorization: generateValidRequestAuthorizationHeader(userId),
+        },
+        method: 'DELETE',
+        url: `/api/sessions/${sessionId}`,
+      };
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+
+      const session = await knex('sessions').where({ id: sessionId }).first();
+
+      expect(response.statusCode).to.equal(204);
+      expect(session).to.be.undefined;
+    });
+  });
+});
